cloud/skydive-safety-monitor/requirements.txt
flask==3.0.3
gunicorn==23.0.0
google-cloud-storage==2.17.0
opencv-python-headless==4.10.0.84
numpy==1.26.4